untitled/requirements.txt
django==4.1.4
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.

djangorestframework~=3.14.0